requirements.txt
fastapi
uvicorn
requests
tensorflow
pillow
numpy